+++ src/components/Profile/Title.tsx
@@ -1,9 +1,9 @@
 import { useMotionValueEvent, useScroll } from 'framer-motion';
-import { useRouter } from 'next/navigation.js';
 import { useState } from 'react';
 
 import ComeBackIcon from '@/assets/comeback.svg';
 import { FollowButton } from '@/components/Profile/FollowButton.js';
+import { useComeBack } from '@/hooks/useComeback.js';
 import { useIsMedium } from '@/hooks/useMediaQuery.js';
 import type { Profile } from '@/providers/types/SocialMedia.js';
 
@@ -14,7 +14,6 @@ interface TitleProps {
 export function Title({ profile }: TitleProps) {
     const [reached, setReached] = useState(false);
 
-    const router = useRouter();
     const { scrollY } = useScroll();
     const isMedium = useIsMedium();
 
@@ -22,10 +21,12 @@ export function Title({ profile }: TitleProps) {
         setReached(value > 48);
     });
 
+    const comeback = useComeBack();
+
     return (
         <div className="sticky top-0 z-50 flex h-[60px] items-center justify-between bg-primaryBottom px-4">
             <div className="flex items-center gap-7">
-                <ComeBackIcon className=" cursor-pointer text-lightMain" onClick={() => router.back()} />
+                <ComeBackIcon className=" cursor-pointer text-lightMain" onClick={comeback} />
                 <span className=" text-xl font-black text-lightMain">{profile.displayName ?? '-'}</span>
             </div>
